expandin on backing data type for easier conversion to and from various backing types
